float instead of int

+++ web/app/themes/bonsai/includes/functions/adt-person-functions.php
@@ -80,7 +80,7 @@ function adt_get_person_footprint()
     wp_send_json_success($data);
 }
 
-function get_total_value(array $fdemand_categories, string $country, string $act_code, int|string $version) : int {
+function get_total_value(array $fdemand_categories, string $country, string $act_code, int|string $version) : float {
     global $SEPARATOR;
     $total = 0;
     foreach ($fdemand_categories as $cat){
